requirements.txt
streamlit
langchain
langchain-community
faiss-cpu
transformers
torch
pypdf
sentence-transformers
langchain-huggingface
google-generativeai
python-dotenv
tiktoken
reportlab
pandas
numpy
langchain-text-splitters
ruff
pytest
